+++ src/.internal/charts/hierarchy/LinkedHierarchy.ts
@@ -214,7 +214,7 @@ export abstract class LinkedHierarchy extends Hierarchy {
 		super._processDataItem(dataItem);
 
 		const parentDataItem = dataItem.get("parent");
-		if (parentDataItem) {
+		if (parentDataItem && parentDataItem.get("depth") >= this.get("topDepth")) {
 			const link = this.linkDataItems(parentDataItem, dataItem);
 			dataItem.setRaw("parentLink", link);
 		}